requirements.txt
cassandra-driver==3.24.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-API.yandex==0.6.2.1
geomet==0.2.1.post1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
Naked==0.1.31
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
status==0.2.5
urllib3==1.25.10
Werkzeug==1.0.1
